+++ DlssUpdater/GameLibrary/ILibrary.cs
@@ -15,6 +15,7 @@ public enum LibraryType
     Ubisoft,
     EpicGames,
     GOG,
+    Xbox,
     // TODO: More launchers
 }
 
@@ -33,6 +34,7 @@ public interface ILibrary
             LibraryType.Ubisoft => new UbisoftConnectLibrary(config, logger),
             LibraryType.GOG => new GOGLibrary(config, logger),
             LibraryType.EpicGames => new EpicGamesLibrary(config, logger),
+            LibraryType.Xbox => new XboxLibrary(config, logger),
             _ => throw new InvalidEnumArgumentException(nameof(config.LibraryType), (int)config.LibraryType, typeof(LibraryType))
         };
     }
@@ -46,6 +48,7 @@ public interface ILibrary
             LibraryType.Ubisoft => "Ubisoft Connect",
             LibraryType.GOG => "GOG",
             LibraryType.EpicGames => "Epic Games",
+            LibraryType.Xbox => "Xbox",
             _ => throw new InvalidEnumArgumentException(nameof(type), (int)type, typeof(LibraryType))
         };
     }